+++ resources/views/admin/supervisions/edit.blade.php
@@ -28,11 +28,19 @@
                     {!! Form::close() !!}
                     <br>
                     
-                    {!! Form::open(['route'=> ['supervisions.destroy', $supervision->id ], 'method' => 'DELETE']) !!}
-                                        <button type="submit" class="btn btn-danger pull-right" >Eliminar
+                    <br>
+                    @if($supervision->requests()->count() == 0)
+                         {!! Form::open(['route'=> ['supervisions.destroy', $supervision->id ], 'method' => 'DELETE']) !!}
+                                        <button id="deleteRequestsupervisionButton" type="submit" class="btn btn-danger pull-right">Eliminar
+                                        </button>
+                        {!! Form::close() !!}
+                    @else
+                        {!! Form::open(['route'=> ['supervisions.destroy', $supervision->id ], 'method' => 'DELETE']) !!}
+                                        <button id="deleteRequestsupervisionButton" type="submit" class="btn btn-danger pull-right" disabled>Eliminar
                                         </button>
                                        
                     {!! Form::close() !!}
+                    @endif
                
                     
                 </div><!--.panel-body-->
